celery-worker/requirements.txt
celery
redis
psycopg2-binary
flask